Change usage of GCMouse API to fix #1
When not running from Xcode, the mouse pointer would not move until the app had been minimized and opened again. I filed a Feedback with Apple and they responded by pointing out the Notification GCMouseDidBecomeCurrent, which should be used to wait for the GCMouse availability. The previous implementation simply checked for a GCMouse.current in viewDidAppear.