Update project to include iPhone in supported devices #2
This adds the iPhone family to target devices for the app in Xcode, and removes the "iPad" mention in the UI.
Note that Apple does not offer official pointer support on iPhone, so mice must be paired through Accessibility settings. See #2 for more discussion of the topic. iPad will remain the primary targeted device for this project.